requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.0
Django==3.2.4
djangorestframework==3.12.4
environ==1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.0
psycopg2-binary==2.9.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
qrcode==7.2
reportlab==3.5.68
segno==1.3.3
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.2
webencodings==0.5.1
xhtml2pdf==0.2.5
